# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateEndpointRequest:
    properties:
      expected_status:
        maximum: 599
        minimum: 100
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      url:
        type: string
    required:
    - expected_status
    - name
    - url
    type: object
...
  dto.EndpointListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Endpoint'
        type: array
      success:
        type: boolean
    type: object
  dto.EndpointStatsResponse:
    properties:
      data:
        $ref: '#/definitions/stats.EndpointStats'
      success:
        type: boolean
    type: object
  dto.EndpointSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/models.Endpoint'
      success:
        type: boolean
    type: object
...
  dto.IncidentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.IncidentResponse'
        type: array
      success:
        type: boolean
    type: object
  dto.IncidentResponse:
    properties:
      endpoint_id:
        type: integer
      id:
        type: integer
      is_resolved:
        type: boolean
      resolved_at:
        type: string
      started_at:
        type: string
    type: object
  dto.IncidentSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.IncidentResponse'
      success:
        type: boolean
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.MonitoringResponse:
    properties:
      endpoint_id:
        type: integer
      monitoring_started_at:
        type: string
    type: object
  dto.MonitoringSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.MonitoringResponse'
      success:
        type: boolean
    type: object
...
  dto.UpdateEndpointRequest:
    properties:
      expected_status:
        maximum: 599
        minimum: 100
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      url:
        type: string
    required:
    - expected_status
    - name
    - url
    type: object
...
  models.Endpoint:
    properties:
      checkInterval:
        type: integer
      createdAt:
        type: string
      expectedStatus:
        type: integer
      id:
        type: integer
      lastCheckedAt:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      url:
        type: string
      userID:
        type: integer
    type: object
  stats.EndpointStats:
    properties:
      average_latency:
        type: number
      success_rate:
        type: number
      total_checks:
        type: integer
    type: object
  utils.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  utils.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend API for monitoring HTTP endpoints.
  title: API Performance Observatory API
  version: "1.0"
paths:
...
  /endpoints:
    get:
      consumes:
      - application/json
      description: Returns all monitored endpoints.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EndpointListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List endpoints
      tags:
      - Endpoints
    post:
      consumes:
      - application/json
      description: Creates a new endpoint to monitor.
      parameters:
      - description: Endpoint details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEndpointRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EndpointSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create endpoint
      tags:
      - Endpoints
  /endpoints/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an endpoint.
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete endpoint
      tags:
      - Endpoints
    get:
      consumes:
      - application/json
      description: Returns one endpoint.
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EndpointSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get endpoint
      tags:
      - Endpoints
    put:
      consumes:
      - application/json
      description: Updates an existing endpoint.
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated endpoint
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEndpointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EndpointSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update endpoint
      tags:
      - Endpoints
  /endpoints/{id}/incidents:
    get:
      consumes:
      - application/json
      description: Returns the currently active incident for an endpoint.
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IncidentSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get active incident for endpoint
      tags:
      - Incidents
  /endpoints/{id}/monitoring:
    get:
      consumes:
      - application/json
      description: Returns monitoring information for an endpoint.
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MonitoringSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get monitoring information
      tags:
      - Monitoring
  /endpoints/{id}/stats:
    get:
      consumes:
      - application/json
      description: Returns aggregated statistics calculated from all health checks
        for an endpoint.
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EndpointStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get endpoint statistics
      tags:
      - Statistics
...
  /incidents:
    get:
      consumes:
      - application/json
      description: Returns all recorded incidents.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IncidentListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List incidents
      tags:
      - Incidents
  /incidents/{id}:
    get:
      consumes:
      - application/json
      description: Returns an incident by ID.
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IncidentSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get incident
      tags:
      - Incidents
  /incidents/active:
    get:
      consumes:
      - application/json
      description: Returns all unresolved incidents.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IncidentListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List active incidents
      tags:
      - Incidents
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"